+++ ext/polars/src/conversion/mod.rs
@@ -41,18 +41,6 @@ use crate::{
     RbDataFrame, RbExpr, RbLazyFrame, RbPolarsErr, RbResult, RbSeries, RbTypeError, RbValueError,
 };
 
-pub(crate) fn slice_extract_wrapped<T>(slice: &[Wrap<T>]) -> &[T] {
-    // Safety:
-    // Wrap is transparent.
-    unsafe { std::mem::transmute(slice) }
-}
-
-pub(crate) fn vec_extract_wrapped<T>(buf: Vec<Wrap<T>>) -> Vec<T> {
-    // Safety:
-    // Wrap is transparent.
-    unsafe { std::mem::transmute(buf) }
-}
-
 #[repr(transparent)]
 pub struct Wrap<T>(pub T);