+++ lib/polars/data_frame.rb
@@ -6448,7 +6448,7 @@ module Polars
     #   #         {5,"five"}
     #   # ]
     def to_struct(name = "")
-      Utils.wrap_s(_df.to_struct(name))
+      Utils.wrap_s(_df.to_struct(name, []))
     end
 
     # Decompose a struct into its fields.